+++ Assets/3D Assets/Rudolph.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f6929b0340a25514dbda58bf3bf706ef
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
