+++ Assets/3D Assets/Rudolph/Materials/0.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fd7a79e5e194fb47a5cbed54cbb4569
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
