+++ Assets/3D Assets/Rudolph/Rudolph IDle.fbm.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 137859b8cb9afba4ebce6575b10d361b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
