+++ Assets/3D Assets/Rudolph/Rudolph.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c1152412ccfabd4e8c1eb4a23eb9cb0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
