+++ Assets/3D Assets/Rudolph/Rudolph.fbm.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1ad8eab86264e114fa6ff01472a93baf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
