+++ Assets/Scripts/Managers/WaveData 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: edc0b016a13f4d6448fe7acb7baab049
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
