+++ Assets/Scripts/Managers/WaveData 2.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d813cc86e16e564d88f2a2fc8c0b087, type: 3}
+  m_Name: WaveData 2
+  m_EditorClassIdentifier: 
+  enemyPrefab: {fileID: 4040545918539713344, guid: f36d7a4bcf190fc4a9a811acc4f91b37, type: 3}
+  enemyCount: 10
+  spawnInterval: 1
+  cooldownTime: 5
