+++ Assets/Scripts/Managers/WaveData 2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c43d02a176e41874d81b00613da776ae
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
