+++ Assets/Scripts/Managers/WaveData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5d813cc86e16e564d88f2a2fc8c0b087
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
